+++ CHANGELOG.md
@@ -1,3 +1,24 @@
+## [1.30.5](https://github.com/openwhyd/openwhyd/compare/v1.30.4...v1.30.5) (2020-03-22)
+
+
+### Bug Fixes
+
+* **tests:** Migrate acceptance tests from Webdriver to Cypress ([#301](https://github.com/openwhyd/openwhyd/issues/301)) ([2f6de36](https://github.com/openwhyd/openwhyd/commit/2f6de36)), closes [#199](https://github.com/openwhyd/openwhyd/issues/199)
+
+## [1.30.4](https://github.com/openwhyd/openwhyd/compare/v1.30.3...v1.30.4) (2020-03-22)
+
+
+### Bug Fixes
+
+* **tests:** create dummy posts from cypress tests instead of from initdb_testing ([#304](https://github.com/openwhyd/openwhyd/issues/304)) ([e37a0d0](https://github.com/openwhyd/openwhyd/commit/e37a0d0)), closes [#199](https://github.com/openwhyd/openwhyd/issues/199)
+
+## [1.30.3](https://github.com/openwhyd/openwhyd/compare/v1.30.2...v1.30.3) (2020-03-22)
+
+
+### Bug Fixes
+
+* **tests:** Reset database between each Cypress tests ([#302](https://github.com/openwhyd/openwhyd/issues/302)) ([ced6038](https://github.com/openwhyd/openwhyd/commit/ced6038)), closes [#301](https://github.com/openwhyd/openwhyd/issues/301) [#199](https://github.com/openwhyd/openwhyd/issues/199)
+
 ## [1.30.2](https://github.com/openwhyd/openwhyd/compare/v1.30.1...v1.30.2) (2020-03-22)